msc-4-hands/requirements.freeze.txt
absl-py==1.0.0
anyio==3.5.0
apparmor==3.0.4
appdirs==1.4.4
apsw==3.36.0.post1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asciidoc==10.1.3
asn1crypto==1.4.0
astor==0.8.1
astroid==2.9.3
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blessed==1.19.1
brotlicffi==1.0.9.2
btrfsutil==5.16.2
CacheControl==0.12.6
cached-property==1.5.2
cachetools==5.0.0
caffeine-ng==3.5.1.post27+g1e5febf.d20211213
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
css-parser==1.0.6
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
dnspython==2.2.0
docker==5.0.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
dotmap==1.3.26
enlighten==1.10.2
entrypoints==0.4
evdev==1.4.0
ewmh==0.1.6
executing==0.8.2
feedparser==6.0.2
flatbuffers==2.0.6
fonttools==4.29.1
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
gpg==1.17.0
grpcio==1.43.2
gufw==21.4.0
h5py==3.6.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
icecream==2.1.2
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.8.1
ipykernel==6.9.1
ipython==8.1.0
ipython_genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras @ file:///build/python-keras/src/keras-2.8.0-py2.py3-none-any.whl
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lensfun==0.3.95
LibAppArmor==3.0.4
libfdt==1.6.1
libtorrent===2.0.5-build-libtorrent-rasterbar-src-libtorrent-rasterbar-2.0.5-bindings-python
lit==13.0.1.dev0
lxml==4.7.1
manjaro-sdk==0.1
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mechanize==0.4.6
mistune==0.8.4
more-itertools==8.10.0
msgpack==1.0.3
multivolumefile==0.2.3
mypy-extensions==0.4.3
nbclassic==0.3.6
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
netsnmp-python==1.0a1
notebook==6.4.8
notebook-shim==0.1.0
npyscreen==4.10.5
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.2
pathspec==0.9.0
pathtools==0.1.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
ply==3.11
prefixed==0.3.2
progress==1.6
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pulsectl==22.1.3
pure-eval==0.2.2
pwquality==1.4.4
py7zr==0.16.2
pyalpm==0.10.6
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.1
pycairo==1.20.1
pychm==0.8.6
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.10.1
pycuda==2021.1
pycups==2.0.1
pycurl==7.44.1
pydot==1.4.2
Pygments==2.11.2
PyGObject==3.42.0
pylint==2.12.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.0
pyppmd==0.17.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyrsistent==0.18.1
PySide6==6.2.3
pysmbc==1.0.23
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-xlib==0.31
pytools==2022.1
pytz==2021.3
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
pyzstd==0.15.0
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
resolvelib==0.5.5
retrying==1.3.3
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.5.7
setproctitle==1.2.2
sgmllib3k==1.0.0
shiboken6==6.2.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
TBB==0.2
team==1.0
tensorboard==2.8.0
tensorboard-data-server @ file:///tmp/tensorboard_data_server-0.7.0a0-py3-none-manylinux2010_x86_64.whl
tensorboard-plugin-wit @ file:///build/python-tensorboard_plugin_wit/src/tensorboard_plugin_wit-1.8.1/wit-pip/release/dist/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow-estimator @ file:///build/python-tensorflow-estimator/src/estimator_pip/tensorflow_estimator-2.7.0-py2.py3-none-any.whl
tensorflow-gpu @ file:///build/tensorflow/src/tmpoptcuda/tensorflow_gpu-2.8.0-cp310-cp310-linux_x86_64.whl
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
texttable==1.6.4
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.0
torbrowser-launcher==0.3.5
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.0
ufw==0.36.1
unrardll==0.1.4
urllib3==1.26.8
wandb==0.12.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yaspin==2.1.0
zeroconf==0.38.2
zipp==3.7.0
